BACKLOG-21127: Expand Structured view to display selected items
